Moved InputEventStore into UI class. Updated InputEventHandler to a pure interface, moving implementation details to the inheritors
