+++ CHANGELOG.md
@@ -4,8 +4,6 @@
 
 - More UI widgets (Input, Checkbox, Slider)
 - Text widget spanning to allow different text attributes for different segments of text
-- Minimum size computation in widgets
-- Sized widget minimum size configuration
 - Dynamic lighting
 - Networked multiplayer
 - Save/Load system
